Factored out activity tests
